refactor: move fetch utils into own folder
