+++ utils/fetch/index.js
@@ -1,0 +1,5 @@
+const latestVersion = require('./latest-version');
+const packageMeta = require('./package-meta');
+const sendCommand = require('./send-command');
+
+module.exports = { latestVersion, packageMeta, sendCommand };
