+++ utils/index.js
@@ -1,17 +1,7 @@
 'use strict';
 
-const sendCommand = require('./send-command');
 const logger = require('./logger');
-const fetchLatestVersion = require('./fetch-latest-version');
-const fetchPackageMeta = require('./fetch-package-meta');
 const getDefaults = require('./get-defaults');
 const getCWD = require('./get-cwd');
 
-module.exports = {
-    sendCommand,
-    logger,
-    fetchLatestVersion,
-    fetchPackageMeta,
-    getDefaults,
-    getCWD,
-};
+module.exports = { logger, getDefaults, getCWD };
